requirements.txt
requests>=2.28
paho-mqtt>=2.1
